Introduce OCR capabilities
This commit adds an OCR (Optical Character Recognition) module to the
project.

The module enables the extraction of text blocks from a frame using the
PSM (Page Segmentation Mode) and two arrays representing the lower and
upper bounds of the content color.

Additionally, while the module offers its own method for enhancing
images to extract text, it might also use a user-defined function.